http: removed date caching to fix delay in exit
Without this patch, there can be up to a second delay between the time
the server is closed and the time the node process is exited, because
the utcDate() function in OutgoingMessage was doing a setTimeout() to
update the cached date/time.

This won't matter as much for production servers, but should have a
bigger impact on the tests.

If it turns out that the date/time caching was not a premature
optimization and should be retained, an `.unref()` should be added to
the `setTimeout()` timer.

I worked on this during nodeconf2013 -- @isaacs nailed down the root
problem and @piscisaureus helped quite a bit.