http: Send GET bodies w/out Content-Length chunked
If a GET with a body is being sent and it is in implicit headers mode,
the body will be sent chunked by default (if a Content-Length is
not specified). Otherwise it would result in a request with a body and
no Content-Length header, which is invalid and causes a socket hang up
error.

The HTTP 1.1 spec allows sending a body with a GET request, but the
server isn't allowed to attach meaning to it because GET requests are
idempotent.