crypto: Explicitly check for RSA keys
I'm not entirely sure why, but if you provide a Certificate containing an ECC key for SSL Client Authentication and call getPeerCertificate() against it, the SSL connection will reset in some cases upon attempting to send data. This fix does an explicit check for an RSA key before performing the EVP_PKEY_get1_RSA() call. That corrects the problem.