net: family option in net.connect
There is issue with dns.lookup in net module. If lookup fails with IPv4
this code will throw an error and there is no way to specify another
stack version to fix this. In general it will throw ENETUNREACH error
when IPv4 address is not available, but IPv6 address is.

This patch appends family option to specify stack verison in net.connect
and net.createConnection options. With this options programmers can
write http.Agent with overridden createSocket which will catch
ENETUNREACH and retry with another stack version.