src: add support to toggle debug mode
Reason: Developers need to switch in and out of debug mode on
production hosts that are running node applications.

Earlier, SIGUSR1 signal would put node into debug mode
but there was no way to come out of it. With this change you can use
SIGUSR1 to come out of debug mode if debugger is already running.