src: add accessor method for debugger_running flag
Reason: If a developer wants to implement a module that needs
to know debug mode status.

Example: We use a module process-watcher which monitors nodejs
processes for endless loops and other abnormal behaviours and
kills it. We would not want it to kill a worker if it in debug
mode. To ensure that we have a cpp module, monitr, that sends
the debug mode status to process-watcher.