Simplify the logic in vcbuild.bat
While adding support for Visual Studio 2013, I noticed duplicate
logic for VS detection in gyp (MSVS*.py), nodejs (vcbuild.bat),
and node-gyp (lib/configure.js).  The detection logic in gyp is
sufficient for our needs, so I am simplifing the logic in nodejs
and node-gyp.

This change requires node to be built from a VS Command Prompt
as it checks for the existence of VS1x0COMNTOOLS and relies on
msbuild being on the path.

This change also updates the README to specify using a command
prompt, note the support of VS 2013, and add hyperlinks to the
downloads.