debugger: deprecate built-in debugger
Reduce the maintenance burden a little by deprecating the built-in
debugger.

Polls in the IRC channel and on the mailing list suggest that not many
people use the debugger or even know it exists.  The few that do are
generally happy to see it get moved into a separate npm module.

This commit takes the first step my deprecating the debugger and
removing its tests.  The tests were very flaky anyway, they won't
be missed.