crypto: don't filter uppercase ciphers/hashes
Stop filtering out the uppercased cipher/hash names in getCiphers() and
getHashes() because some don't have lowercase versions (notably DSA).

OpenSSL's EVP_get_cipherbyname() and EVP_get_digestbyname() functions
are case-sensitive so it's essential that we don't munge the name in
that case (pun unintended.)

We could filter out the uppercase version iff there is a lowercase
version but that makes both the code and the mental model for the
user more complex.