TLS Compression is used for CRIME attacks, not BEAST.
This is a comment change, where it originally says disabling TLS Compression protects against BEAST attack. But in fact, it is the CRIME attack(Compression Ratio Info-leak Made Easy) that makes use of TLS Compression and not BEAST. 

BEAST(Browser Exploit Against SSL/TLS) is an entirely another variant making use of the chosen boundary attack against CBC mode in encryption.

Just making sure, that the exact reason for disabling TLS compression must be made clear and not be misleading with some other attack. Will surely be useful down the years.