cluster: RR domain missing method addAsyncListener
Failing test is provided. I am not familiar enough
with the core C code to provide a fix.

Summary
-------

Using the new round robin (RR) scheduler (which is the default),
if a worker has a domain when it calls server.listen, a
TypeError: Object has no method 'addAsyncListener' is thrown.

```
net.js:1095
    this._handle.addAsyncListener(this.domain._listener);
                 ^
TypeError: Object #<Object> has no method 'addAsyncListener'
    at Server._listen2 (net.js:1095:18)
    at cb (net.js:1139:10)
    at rr (cluster.js:552:5)
    at Worker.<anonymous> (cluster.js:484:9)
    at process.<anonymous> (cluster.js:611:8)
    at process.EventEmitter.emit (events.js:123:20)
    at handleMessage (child_process.js:318:10)
    at Pipe.channel.onread (child_process.js:346:11)
```

Details
-------

The application I was testing node v0.11.9 and v0.11.10 with
creates a domain to encapsulate a postgres DB connection
before calling server.listen.

With the new round robin scheduler as the default this
immediately throws the error.

Forcing cluster to use the old scheduler (which is done
by the OS) works as expected, so the issue must be related
to the handle in round robin workers not having the new
addAsyncLister method and related functionality.

Work around
-----------

The issue can be worked around by forcing the old scheduler
with code or by setting the env variable as follows:

```javascript
 cluster.schedulingPolicy = cluster.SCHED_NONE;
```

OR using the env variable

```bash
env NODE_CLUSTER_SCHED_POLICY="none" node app.js
```