node: pass calling context to create/error
Now that the instance has finished instantiating before create is
called, it is possible to pass the calling context as the first
argument.

The current calling context is stored within scope to access the
asyncQueue attached. This has the added benefit of allowing the context
to be passed to the error callback.

Currently this change will cause a SEGFAULT if properties are inspected
from specific contexts!!!