module: strip leading lines when -x switch set
Factor the code that removes '#!' into a function,
Module._removePrologue.

Module._removePrologue removes only the first #! line unless
BOTH a) this is the main module (module.id === '.')
 AND b) the '-x' switch was passed on the command line.

In case both a) and b) are true, Module._removePrologue
removes lines until it finds a line which begins with "#!"
and contains the substring "node".