http: remove path from Request options
Some 3rd request modules uses path property.  For instance:
https://github.com/mikeal/request/blob/master/request.js#L251

It did not matter in the Node 0.10. But in Node 0.11 it leads
to malfunction of the net core module. This is because the path
is considered as a pipe (v0.11.11-release/lib/net.js#L844).

It is not a bug. Although it is an eco-friendly way to get rid of a problem.