domain: remove multiple of the same from stack
If a domain is active when the callback is called, then the domain is
entered. When the user manually enters a domain before the callback is
run then there will be a duplicate copy of the domain in the stack.

Instead remove to the first instance of the domain pushed to the stack.