fs: Improve performance of fs.stat and related functions.
As suggested by @trevnorris in joyent/node#6662 moved the setting of
object properties from C++ to JavaScript. Specifically, remove calls to
stats->Set and instead set properties of the stats object in JavaScript
with a call to a new function fs.Stats.prototype.buildStatsObject.

Note that the returned value from fs.stat changes slightly for non-POSIX
systems. Whereas before the stats object would be missing blocks and
blksize keys, it now has these keys with undefined as the value.