domain: fix parents tracking
We cannot use current stack to keep track of domains chain and their
parents for exceptions handling, we need an additional property to do
so. The reason behind it is an asynchronous code that throws an
exception and so when domain is handling that exception, the stack is
empty and there is no way to get the parent.