buffer: Speed up buffer write benchmarks
The buffer-write benchmarks were themselves so slow that it masked any
change in the effectiveness of the buffer write routines. Prior to this
commit, over 90% of the overhead was due to calling the correct write*
function using buff[fn] rather than the much more optimized
buff.write* function as would usually be done in the real world. I have
never seen any production Node.js code that calls buffer write
functions in the way the benchmark had them.

The use of new Function() will no doubt annoy some dogmatic idiots. The
alternatives are to either write out an assload of functions by hand or
have a useless benchmark. The first function of a benchmark is to be
useful. I'd argue that deleting this benchmark would be better than
reverting it to a state where over 90% of what it measures is the speed
of property lookups.

The last change that is likely to be controversial is changing from
using mod (%) to bitwise AND (&). This will remove the cost of Modding
ints from the benchmark. There is no reason to believe that that cost
is important for this benchmark, so it has been removed.